+++ src/Toml/Semantics/Ordered.hs
@@ -39,10 +39,12 @@ import Data.Foldable (foldl', toList)
 import Data.List (sortOn)
 import Data.Map (Map)
 import Data.Map qualified as Map
+import Data.Text (Text)
+import Data.Text qualified as Text
 import Toml.Parser.Types (Expr(..), Key, Val(ValTable, ValArray))
 
 -- | Summary of the order of the keys in a TOML document.
-newtype TableOrder = TO (Map String KeyOrder)
+newtype TableOrder = TO (Map Text KeyOrder)
 
 -- | Internal type used by 'TableOrder'
 --
@@ -51,14 +53,14 @@ newtype TableOrder = TO (Map String KeyOrder)
 data KeyOrder = KeyOrder !Int TableOrder
 
 -- | Opaque type used by 'projectKey'
-newtype ProjectedKey = PK (Either Int String)
+newtype ProjectedKey = PK (Either Int Text)
     deriving (Eq, Ord)
 
 -- | Generate a projection function for use with 'Toml.Pretty.prettyTomlOrdered'
 projectKey ::
     TableOrder   {- ^ table order                           -} ->
-    [String]     {- ^ table path                            -} ->
-    String       {- ^ key                                   -} ->
+    [Text]       {- ^ table path                            -} ->
+    Text         {- ^ key                                   -} ->
     ProjectedKey {- ^ type suitable for ordering table keys -}
 projectKey (TO to) [] = \k ->
     case Map.lookup k to of
@@ -77,13 +79,13 @@ emptyOrder = TO Map.empty
 extractTableOrder :: [Expr a] -> TableOrder
 extractTableOrder = snd . foldl' addExpr ([], emptyOrder)
 
-addExpr :: ([String], TableOrder) -> Expr a -> ([String], TableOrder)
+addExpr :: ([Text], TableOrder) -> Expr a -> ([Text], TableOrder)
 addExpr (prefix, to) = \case
     TableExpr k      -> let k' = keyPath k in (k', addKey to k')
     ArrayTableExpr k -> let k' = keyPath k in (k', addKey to k')
     KeyValExpr k v   -> (prefix, addVal prefix (addKey to (prefix ++ keyPath k)) v)
 
-addVal :: [String] -> TableOrder -> Val a -> TableOrder
+addVal :: [Text] -> TableOrder -> Val a -> TableOrder
 addVal prefix to lval =
     case lval of
         ValArray _ xs -> foldl' (addVal prefix) to xs
@@ -93,14 +95,14 @@ addVal prefix to lval =
                 addVal k' (addKey acc k') v) to kvs
         _ -> to
 
-addKey :: TableOrder -> [String] -> TableOrder
+addKey :: TableOrder -> [Text] -> TableOrder
 addKey to [] = to
 addKey (TO to) (x:xs) = TO (Map.alter f x to)
     where
         f Nothing = Just (KeyOrder (Map.size to) (addKey emptyOrder xs))
         f (Just (KeyOrder i m)) = Just (KeyOrder i (addKey m xs))
 
-keyPath :: Key a -> [String]
+keyPath :: Key a -> [Text]
 keyPath = map snd . toList
 
 -- | Render a white-space nested representation of the key ordering extracted
@@ -113,7 +115,7 @@ debugTableOrder to = unlines (go 0 to [])
                 (sortOn p (Map.assocs m))
 
         go1 i (k, KeyOrder _ v) z =
-            (replicate (4*i) ' ' ++ k) :
+            (replicate (4*i) ' ' ++ Text.unpack k) :
             go (i+1) v z
 
         p (_, KeyOrder i _) = i