+++ test/ToValueSpec.hs
@@ -1,3 +1,4 @@
+{-# Language OverloadedStrings #-}
 module ToValueSpec where
 
 import Test.Hspec (it, shouldBe, Spec)
@@ -7,10 +8,10 @@ import Toml.ToValue (ToValue(toValue))
 spec :: Spec
 spec =
  do it "converts characters as singleton strings" $
-        toValue '!' `shouldBe` String "!"
+        toValue '!' `shouldBe` Text "!"
 
     it "converts strings normally" $
-        toValue "demo" `shouldBe` String "demo"
+        toValue ("demo" :: String) `shouldBe` Text "demo"
 
     it "converts lists" $
-        toValue [1,2,3::Int] `shouldBe` Array [Integer 1, Integer 2, Integer 3]
+        toValue [1,2,3::Int] `shouldBe` List [Integer 1, Integer 2, Integer 3]